requirements.txt
Django>=3.0
psycopg2-binary>=2.8
djangorestframework==3.14
django-filter==23.1
black==23.10.1